Create Twitter Sentiment Analysis
Sentiment Analysis is the process of ‘computationally’ determining whether a piece of writing is positive, negative or neutral.
We are running it on Python 2.7. It contains Tweepy which is the official python client for Twitter API. Here we use TextBlob, it is the official python library for processing textual data.